chore: restructuring for consistency
